+++ src/vroom_fct.h
@@ -171,18 +171,10 @@ public:
 
     cpp11::sexp x_(x);
 
-    cpp11::integers in(indx);
+    auto idx = get_subset_index(indx);
 
-    auto idx = std::make_shared<std::vector<size_t>>();
-
-    idx->reserve(in.size());
-
-    for (const auto& i : in) {
-      // If there are any NA indices fall back to the default implementation.
-      if (i == NA_INTEGER) {
-        return nullptr;
-      }
-      idx->push_back(i - 1);
+    if (idx == nullptr) {
+      return nullptr;
     }
 
     auto inf = Info(x);