+++ NEWS.md
@@ -1,5 +1,7 @@
 # vroom (development version)
 
+* Column subsetting with double indexes now works again (#257)
+
 # vroom 1.3.0
 
 * The Rcpp dependency has been removed in favor of cpp11.